// src/interfaces/document.interface.ts
export interface IDocument {
  id: string;
  scope: string;
  label: string;
  path: string;
  category: string;
  content: string;
  language: string;
  source: string;
  creator: string;
  status: string;
  company_name: string;
  hr_page: number;
  doc_type: string;
  document_name: string;
  create_date: string;
  published_date: string;
}